+++ R/DTAData-class.R
@@ -1,0 +1,20 @@
+#' @title DTAData Class
+#' @description Handles data from a file.
+#' @import S7
+#'
+#' @param specs A DTAColumnSpecCollection object specifying the column specs.
+#' @param data List. A list of tables to be validated and included in the DTAData object.
+#' @return An object of class DTAData. 
+#
+#' @examples
+#'
+#' \dontrun{
+#' }
+#' @export
+DTAData <- new_class(
+  "DTAData",
+  constructor = function(
+  ) {
+  
+  }
+)