+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2025-09-11
+
+### Added
+
+- Added GitHub Action workflows
+- Added Metadata to DTAData
+
+### Changes
+
+- Moved project to GitHub.com
+- Modified documentation and examples to adhere to R CMD checks for CRAN
+
 ## [0.8.1] - 2025-08-18
 
 ### Added